feat: added filtering for missing tables
