+++ src/composables/useTrafficStats.ts
@@ -11,7 +11,7 @@ export type TrafficRecord = {
 export type TrafficSummary = {
   total_traffic: number
   hourly_traffic: TrafficRecord[]
-  applications: TrafficRecord[]
+  applications?: TrafficRecord[]
   clients: TrafficRecord[]
   remote_hosts?: TrafficRecord[]
   protocols?: TrafficRecord[]
@@ -23,21 +23,30 @@ export function useTrafficStats() {
   const data = ref<TrafficSummary>({
     total_traffic: 0,
     hourly_traffic: [],
-    applications: [],
     clients: []
   })
 
-  function loadData(client?: string, application?: string) {
+  function loadData(client?: string, application?: string, protocol?: string, host?: string) {
     loading.value = true
     const payload: {
       client?: string
-      application?: string
+      section?: string
+      value?: string
     } = {}
     if (client) {
       payload.client = client
     }
     if (application) {
-      payload.application = application
+      payload.section = 'application'
+      payload.value = application
+    }
+    if (protocol) {
+      payload.section = 'protocol'
+      payload.value = protocol
+    }
+    if (host) {
+      payload.section = 'host'
+      payload.value = host
     }
     ubusCall('ns.dpireport', 'summary-v2', payload)
       .then((response: AxiosResponse<TrafficSummary>) => {
